feat: Customize Django admin for CustomUser model and define custom forms
Customize the Django admin interface for the CustomUser model by defining a
custom admin class (CustomUserAdmin) in admin.py. This admin class specifies
the use of custom forms (CustomUserCreationForm and CustomUserChangeForm) for
adding and changing users. Additionally, define CustomUserCreationForm and
CustomUserChangeForm classes in forms.py, tailored for the CustomUser model.